feat(magda) translate panel style
