+++ src/assets/style/components/_forms.scss
@@ -142,15 +142,24 @@
 
 
 
-form.cms {
+form.cms, form.cms-translate {
   flex-grow: 1;
   margin-left: 5%;
   
   .main-title {
     text-align: left;
-    margin-bottom: 50px;
+    margin: {
+      top: 10px;
+      bottom: 50px;
+      right: 0;
+      left: 0;
+    };
     font-size: 1.3rem;
-    padding: 8px;
+    padding: 10px;
+    background: white;
+    border-radius: 5px;
+    width: 60%;
+    box-shadow: 0 0 0.5rem $crl-main-grey;
   } 
   .form-container {
     border-radius: 5px;
@@ -218,9 +227,7 @@ form.cms {
       padding: 1%;
       border-radius: 4px;
     }
-  }
-
-  
+  } 
 
   .submit-btn {
     display: block;
@@ -241,59 +248,34 @@ form.cms {
   } 
 }
 
-
-
-
 form.cms-translate {
-  display: block;
-  margin: 0 auto;
-  max-width: 900px;
-  .translate-container {
-    margin-top: 50px;
+  .form-control {
+    min-height: 230px;
+    display: block;
+    label {
+      display: block;
+      width: 100%;
+    }
+    textarea, input {
+      background-color: white;
+      width: 95%;
+      border: 1px solid $crl-panel-dark-grey;
+      padding: 1%;
+      border-radius: 4px;
+    }
+  }
 
-    .form-group {
-      border-top: 1px solid $crl-main-placeholder;
-      padding-top: 20px;
-      margin-top: 40px;
-      input, textarea {
-        width: 100%;
-        border-radius: 5px;
-        border: 1px solid $crl-base-blue;
-        padding: 10px;
-      }
-
-      blockquote {
-        font-style: italic;
-        font-size: 0.9rem;
-
-        .blockquote-footer::before {
-          display: none;
-        }
-      }
-      .card {
-        border-radius: 0px;
-        border: none;
-        border-left: 1px solid $crl-base-blue;
-        margin-bottom: 10px;
-      }
+  .card {
+    background: transparent;
+    .card-body {
+      padding: 5px 15px;
+      max-height: 60px;
+      overflow-y: auto;
+    }
+    footer, p{
+      font-size: 0.7rem;
     }
     
   }
-  .submit-btn {
-    display: block;
-    color: $crl-base;
-    font-weight: bold;
-    text-transform: uppercase;
-    font-size: 0.9rem;
-    background-color: $crl-base-blue;
-    border: 1px solid $crl-base-blue;
-    opacity: 1;
-    border-radius: 30px;
-    padding: 10px 35px;
-    margin: 25px auto;
-    cursor: pointer;
-    &:hover {
-      background-color: rgba($crl-base-blue, 0.9);
-    }
-  }
+  
 }
